% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventskalenData.R
\docType{data}
\name{adventskalenData}
\alias{adventskalenData}
\title{Der Inhalt der Türchen}
\format{
verschlüsselte Liste
}
\usage{
data(adventskalenData)
}
\description{
Der Inhalt des Datensatzes hat niemanden zu interessieren. Zugriff nur über das Öffnen von Türen!
}
\details{
Werden nicht verraten.
}
\keyword{datasets}
