% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventskalendR.R
\name{adventskalendR}
\alias{adventskalendR}
\title{Öffnen von Adventskalender-Türchen leicht gemacht}
\usage{
adventskalendR(datum = today())
}
\arguments{
\item{datum}{Standardmäßig das aktuelle Datum. Es können aber auch zwei Datumsangaben als Character angegeben werden, die als Zeitraum interpretiert werden. ACHTUNG: Datumsangaben aus der Zukunft werden gemeldet!}
}
\description{
Mit diese Funktion wird der Adventskalender geöffnet. Natürlich dürfen auch Türen aus der Vergangenheit geöffnet werden. Aber hüte dich davor Türchen zu früh zu öffnen.
}
\examples{
adventskalendR()
adventskalendR(datum = c("2021-12-01", "2021-12-03"))
adventskalendR(datum = c("2021-12-01", as.character(today())))
adventskalendR(datum = "alle")
}
